style: rename files, move contents of patreon_common into patreon/scanner
