--- patreon/patreon_common.py
@@ -1,5 +1,0 @@
-"""
-Stores data shared by files in the patreon module
-"""
-
-baseURL = "https://www.patreon.com"
